comments and my fixes
before getting reponse to memory size and terminal width questions, x and y needed to be
initialized. otherwise they were left at values set by unrelated code and then used

in parse input line, the code to skip to the end of a token string after deciding it wasnt a match
was using MATHTBL+28+1 which I have no idea what that was supposed to mean.
and after finding the end of the token string, needed to iny so that y pointed to
begining of next token string before going back to string compare